% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackingFunctions.R
\name{cleanDataGBM}
\alias{cleanDataGBM}
\title{Clean GBM Data}
\usage{
cleanDataGBM(trainset, gbm_vars, gbm_factors = NULL, testset = NULL,
  new_vars_gbm = NULL)
}
\arguments{
\item{trainset}{The training dataset}

\item{gbm_vars}{Predictor variables to use in the GBM model}

\item{gbm_factors}{Which predictor variables need to be converted to factors}

\item{testset}{Optional test set; included so that both cleaned versions will have the same dimensions}

\item{new_vars_gbm}{Optional variable names to include as input}
}
\value{
A list, containing cleaned versions of the trainset and testset
}
\description{
Cleans the training data for xgboost, which requires a model matrix instead of a formula
}
