% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackingFunctions.R
\name{cleanDataLASSO}
\alias{cleanDataLASSO}
\title{Clean LASSO data}
\usage{
cleanDataLASSO(trainset, lasso_vars, lasso_factors = NULL,
  testset = NULL, new_vars_lasso = NULL)
}
\arguments{
\item{trainset}{The training dataset}

\item{lasso_vars}{Predictor variables to use in the LASSO model}

\item{lasso_factors}{Which predictor variables need to be converted to factors}

\item{testset}{Optional test set; included so that both cleaned versions will have the same dimensions}

\item{new_vars_lasso}{Optional variable names to include as input}
}
\value{
A list, containing cleaned versions of the trainset and testset
}
\description{
Cleans the training data for glmnet, which requires a model matrix instead of a formula
}
