% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{education}
\alias{education}
\title{Marginal Frequency Distribution}
\format{A data frame with 6 rows and 3 variables:
\describe{
  \item{unit}{subnational unit ID}
  \item{education}{a categorical variable denoting education level}
  \item{freq}{the number of persons}
  ...
}}
\source{
Example dataset by Joe Ornstein.
}
\usage{
education
}
\description{
A dataset containing an example poststratification frame
}
\keyword{datasets}
