% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackingFunctions.R
\name{getHillClimbWeights}
\alias{getHillClimbWeights}
\title{Hill Climbing Algorithm}
\usage{
getHillClimbWeights(Y, M, verbose = T)
}
\arguments{
\item{Y}{The outcome variable}

\item{M}{An n x k matrix of out-of-fold predictions of Y, where n is the length of Y, and k is th number of models}
}
\value{
The optimal stacking weights
}
\description{
Hill Climbing solution to stacking problem from Caruana et al (2004); use for minimizing LogLoss
}
