% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackingFunctions.R
\name{getStackWeights}
\alias{getStackWeights}
\title{Generate Stack Weights}
\usage{
getStackWeights(trainset, binaryDepVar = F, hlmFormula, lasso_vars,
  lasso_factors = NULL, forestFormula, mns_best = NULL, knnFormula,
  k_best, gbm_vars, gbm_factors = NULL, gbm_params, gbm_tune,
  nfolds = 5)
}
\arguments{
\item{trainset}{The training dataset}

\item{binaryDepVar}{A boolean denoting whether the outcome variable is binary or continuous. Used to specify tuning parameters for machine learning methods.}

\item{hlmFormula}{A formula object for the hierarchical linear model}

\item{lasso_vars}{The predictor variables to include in the LASSO model}

\item{lasso_factors}{Which predictor variables need to be converted to factors before estimating the LASSO model}

\item{forestFormula}{A formula object for the random forest model}

\item{mns_best}{Optional min node size parameter for random forest model}

\item{knnFormula}{A formula object for the KNN model}

\item{k_best}{Optimal k parameter for KNN model}

\item{gbm_vars}{The predictor variables to include in GBM model}

\item{gbm_factors}{Which predictor variables need to be converted to factors before estimating the GBM}

\item{gbm_params}{List of parameters for the GBM model}

\item{gbm_tune}{xgb.tune object, containing the $best.iteration}

\item{nfolds}{Number of cross-validation folds}
}
\value{
The optimal stacking weights (HLM, LASSO, KNN, Random Forest, GBM)
}
\description{
Main Stacking Procedure: Construct k folds and generate predictions from base models
}
\note{
trainset must include a variable called 'y', denoting the outcome
}
