% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststratificationFunctions.R
\name{getSyntheticPSFrame}
\alias{getSyntheticPSFrame}
\title{Generate a synthetic poststratification frame}
\usage{
getSyntheticPSFrame(PSFrame1, PSFrame2)
}
\arguments{
\item{PSFrame1}{A dataframe containing the empirical frequency for each demographic group within each subnational unit}

\item{PSFrame}{A dataframe containing the empirical frequency for each demographic group within each subnational unit}
}
\value{
A synthetic poststratification frame computed from the marginal frequencies in PSFrame1 and PSFrame2
}
\description{
Returns a synthetic frequency distribution derived from the product of marginal distributions. in two input PSFrames. This can be repeated recursively to include more features.
}
