% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackingFunctions.R
\name{LogLossBinary}
\alias{LogLossBinary}
\title{Compute Log Loss}
\usage{
LogLossBinary(actual, predicted, eps = 1e-15)
}
\arguments{
\item{actual}{A vector of actual values}

\item{predicted}{A vector of predicted values}

\item{eps}{A very small number}
}
\value{
Log Loss
}
\description{
Computes the log loss measure, given predicted and actual values
}
