% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststratificationFunctions.R
\name{poststratify}
\alias{poststratify}
\title{Generate poststratified estimates}
\usage{
poststratify(pred, PSFrame)
}
\arguments{
\item{pred}{A vector of predicted values for each row in PSFrame}

\item{PSFrame}{A dataframe containing the empirical frequency of each unique demographic group within a subnational unit}
}
\value{
A dataframe containing the unit IDs and poststratified estimates
}
\description{
Generate poststratified estimates
}
\note{
PSFrame must have 'unit' variable denoting the unit ID, and 'freq' variable denoting the frequency within each cell

length(pred) must equal nrow(PSFrame)
}
