% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.srp.c}
\alias{predict.srp.c}
\title{The Smooth-Rough Partition model prediction}
\usage{
\method{predict}{srp.c}(object, x, ...)
}
\arguments{
\item{object}{An object of class 'srp.c', returned by \code{\link{srp.c}}.}

\item{x}{A new matrix you wish to fit Smooth-Rough Partition model. The dimension of row is the number of covariates.}

\item{...}{Further parameters that can be passed to \code{\link{predict.srp.c}}.}
}
\value{
\item{yhat}{The vector of predicted values.}
}
\description{
This function performs the predictions from the results of Smooth-Rough Partition fitting.
}
\details{
The Smooth-Rough Partition model is described in "Regularised forecasting via smooth-rough partitioning of the regression coefficients", H. Maeng and P. Fryzlewicz (2018), preprint.
}
\seealso{
\code{\link{sic.c}}, \code{\link{srp.c}}
}
\author{
Hyeyoung Maeng, \email{h.maeng@lse.ac.uk}
}
