% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{race}
\alias{race}
\title{Marginal Frequency Distribution}
\format{A data frame with 8 rows and 3 variables:
\describe{
  \item{unit}{subnational unit ID}
  \item{race}{a categorical variable denoting race}
  \item{freq}{the number of persons}
  ...
}}
\source{
Example dataset by Joe Ornstein.
}
\usage{
race
}
\description{
A dataset containing an example poststratification frame
}
\keyword{datasets}
