% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackingFunctions.R
\name{regress.func}
\alias{regress.func}
\title{Quadratic Programming Algorithm}
\usage{
regress.func(Y, preds.var)
}
\arguments{
\item{Y}{The outcome variable}

\item{preds.var}{An n x k matrix of out-of-fold predictions of Y, where n is the length of Y, and k is th number of models}
}
\value{
The optimal stacking weights
}
\description{
Quadratic programming solution to stacking problem from Grimmer, Messing & Westwood (2017)
}
