% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sex}
\alias{sex}
\title{Marginal Frequency Distribution}
\format{A data frame with 4 rows and 3 variables:
\describe{
  \item{unit}{subnational unit ID}
  \item{sex}{a categorical variable denoting gender}
  \item{freq}{the number of persons}
  ...
}}
\source{
Example dataset by Joe Ornstein.
}
\usage{
sex
}
\description{
A dataset containing an example poststratification frame
}
\keyword{datasets}
