% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sic_c.R
\name{sic.c}
\alias{sic.c}
\title{Optimisation of the Smooth-Rough Partition model}
\usage{
sic.c(x.basis = x.basis, B.basis = B.basis, x = x, y = y,
  cf0 = cf0, maxq = maxq, fixedq = F, L = L, inisp = inisp)
}
\arguments{
\item{x.basis}{The b-spline basis defined for interpolated x in \code{\link{srp.c}}}

\item{B.basis}{The b-spline basis defined for constrained regression coefficient.}

\item{x}{The design matrix used in \code{\link{srp.c}}.}

\item{y}{The response variable used in \code{\link{srp.c}}.}

\item{cf0}{The coefficient matrix obtained by natural cubic spline interpolation of x in \code{\link{ncs}}.}

\item{maxq}{The maximum number of unconstrained parameters if \code{fixedq} is FALSE. Otherwise, it is considered as a unique number of unconstrained parameters.}

\item{fixedq}{If TRUE, \code{maxq} is considered as a fixed number of unconstrained parameters and if FALSE, \code{maxq} is a maximum and a sequence of possible values are investigated to select the optimal.}

\item{L}{The dimension of b-spline expansion for constrained parameters used in \code{\link{srp.c}}.}

\item{inisp}{The initial value for the optimisation of tuning parameters in \code{\link{srp.c}}.}
}
\value{
The following components are obtained only when \code{fixedq} is FALSE:
\item{qhat}{The optimal number of unconstrained parameters.}
\item{sicq}{The vector of Schwarz criterion with length \code{maxq} which is computed for the different number of unconstrained parameters.} The following components are obtained only when \code{fixedq} is TRUE:
\item{muhat}{The estimator of constant parameter.}
\item{bhat}{The vector of evaluated constrained functional regression coefficient.}
\item{ahat}{The vector of unconstrained regression coefficient estimators.}
\item{etahat}{The vector containing both \code{bhat} and \code{ahat} with unevaluated form.}
\item{yhat}{The vector of estimated response variable.}
\item{sp}{The vector of two tuning parameters estimated by minimising generalised cross validation (GCV).}
\item{L}{The number of b-spline bases used for constrained regression parameters.}
}
\description{
This function performs the optimisation of the number of unconstrained regression parameters in Smooth-Rough Partition model by minimising SIC criterion and gives the change-point of smoothness in regression parameters.
}
\details{
Usually only called by \code{\link{srp.c}}.
}
\examples{
library(fda)
x <- matrix(rnorm(10000), ncol=100)
y <- matrix(rnorm(100), ncol=1)
p <- dim(x)[1] + 1
t <- seq(0, 1, length.out=dim(x)[1])*(dim(x)[1])
x.basis <- as.fd(splinefun(t, x[, 1], method="natural"))$basis
B.basis <- create.bspline.basis(rangeval=c(0, dim(x)[1]), norder=4, nbasis=35)
result <- sic.c(x.basis=x.basis, B.basis=B.basis, x=x, y=y, cf0=ncs(x)$cf0, maxq=10, L=35, inisp=1)
plot(result$sicq, type="b")
}
\seealso{
\code{\link{sic.l}}, \code{\link{predict.srp.c}}, \code{\link{srp.c}}
}
\author{
Hyeyoung Maeng, \email{h.maeng@lse.ac.uk}
}
