% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{truebeta}
\alias{truebeta}
\title{A dataset containing true regression coefficients for simulation}
\format{A matrix with 356 rows and 4 variables:
\describe{
  \item{Case 1}{true regression coefficients for case 1}
  \item{Case 2}{true regression coefficients for case 2}
  \item{Case 3}{true regression coefficients for case 3}
  \item{Case 4}{true regression coefficients for case 4}
}}
\usage{
data(truebeta)
}
\description{
True regression coefficients used for simulations in "Regularised forecasting via smooth-rough partitioning of the regression coefficients", H. Maeng and P. Fryzlewicz (2018), preprint.
}
\keyword{datasets}
