% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignetteData}
\alias{vignetteData}
\title{Individual-level Public Opinion and Demographics}
\format{A data frame with 3000 rows and 8 variables:
\describe{
  \item{ID}{respondent ID}
  \item{y}{outcome of interest, continuous}
  \item{x1}{predictor variable, continuous}
  \item{x2}{predictor variable, continuous}
  \item{unit}{subnational unit ID}
  \item{latitude}{subnational unit 'latitude' (stylized)}
  \item{longitude}{subnational unit 'longitude' (stylized)}
  \item{unit_covariate}{unit-level predictor}
  ...
}}
\source{
Ornstein, Joseph T. (2019) "Stacked Regression and Poststratification"
}
\usage{
vignetteData
}
\description{
An example individual-level public opinion dataset
}
\keyword{datasets}
