% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignettePSFrame}
\alias{vignettePSFrame}
\title{Poststratification Frame}
\format{A data frame with 3200 rows and 7 variables:
\describe{
  \item{unit}{subnational unit ID}
  \item{x1}{predictor variable, continuous}
  \item{x2}{predictor variable, continuous}
  \item{freq}{number of persons}
  \item{unit_covariate}{unit-level predictor}
  \item{latitude}{subnational unit 'latitude' (stylized)}
  \item{longitude}{subnational unit 'longitude' (stylized)}
  ...
}}
\source{
Ornstein, Joseph T. (2019) "Stacked Regression and Poststratification"
}
\usage{
vignettePSFrame
}
\description{
A dataset containing an example poststratification frame
}
\keyword{datasets}
