% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignetteTruth}
\alias{vignetteTruth}
\title{True Unit-Level Means}
\format{A data frame with 3200 rows and 7 variables:
\describe{
  \item{unit}{subnational unit ID}
  \item{true_mean}{the true unit-level mean}
  ...
}}
\source{
Ornstein, Joseph T. (2019) "Stacked Regression and Poststratification"
}
\usage{
vignetteTruth
}
\description{
A dataset containing the true unit-level means for the vignette example
}
\keyword{datasets}
