% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srp.R
\name{srp}
\alias{srp}
\title{Stable retrospective power}
\usage{
srp(pvalues, FDR = 0.05, ...)
}
\arguments{
\item{pvalues}{A vector of raw p-values.}

\item{FDR}{A level at which to control the false discovery rate (FDR). Deafults to 0.05.}

\item{...}{Additional arguments passed to \code{\link{qvalue}}.}
}
\value{
A numeric that is the estimated stable retrospective power.
}
\description{
\code{srp} uses p-values and false discovery rate (FDR) to calculate stable retrospective power.
}
\examples{
# import data
library(qvalue)
data("hedenfalk")
pvalues <- hedenfalk$p

# calculate SRP
pw <- srp(pvalues, FDR = 0.05)
pw

}
\seealso{
\code{\link{qvalue}} how q-values and pi0 are calculated and for arguments.
}

