% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srp.R
\name{srp}
\alias{srp}
\title{Stable retrospective power}
\usage{
srp(pvalues, method = "lfdr", FDR = 0.05, ...)
}
\arguments{
\item{pvalues}{A vector of raw p-values.}

\item{FDR}{A level at which to control the false discovery rate (FDR). Deafults to 0.05.}

\item{...}{Additional arguments passed to \code{\link{qvalue}} and \code{\link{pi0est}}.}
}
\value{
A data.frame containing:
 SRP that is the estimated stable retrospective power.
 pi0 propotion of the true non-null effects.
 fp the number of the statistically significant effects that are likely to be false positives.
 rs the number of statistically significant true effects seen in this study that will most likely come up as significant in a replication study.
 ud the number undiscovered results awaiting to be uncovered by replication studies.
}
\description{
\code{srp} uses p-values and false discovery rate (FDR) to calculate stable retrospective power.
}
\examples{
# import data
Test statistics
z <- rnorm(200)
First 40 are have non-zero means
z[1:40] <- z[1:40]+2
True pi0
160/200
Two-sided p-values
p <- 2*pnorm(-abs(z))

# calculate SRP
pw <- srp(pvalues, FDR = 0.05)
pw

}
\seealso{
\code{\link{limma}} and \code{\link{propTrueNull}} how pi0 is calculated and for arguments.
}
